using a more EM-like style of coding
